# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Billing API
  version: 0.2.0
  description: Subscription billing — products, prices, coupons, subscriptions, invoices, and credit notes
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3002
    description: Local billing service. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/tax_rates:
    get:
      summary: List tax rates
      description: Returns a paginated list of tax rates with optional active and inclusive filters.
      x-platform:
        kind: resource
        entity: TaxRate
        affectsLedger: false
      operationId: listTaxRates
      tags:
        - TaxRates
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: active
          description: Filter by active status.
          schema:
            type: string
            enum:
              - "true"
              - "false"
        - in: query
          name: inclusive
          description: Filter by inclusive tax.
          schema:
            type: string
            enum:
              - "true"
              - "false"
      responses:
        "200":
          description: Paginated list of tax rates.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create tax rate
      description: Creates a new tax rate with display name and percentage.
      x-platform:
        kind: resource
        entity: TaxRate
        emitsEvents:
          - tax_rate.created
        affectsLedger: false
      operationId: createTaxRate
      tags:
        - TaxRates
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - display_name
                - percentage
              properties:
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                description:
                  type: string
                  maxLength: 2000
                jurisdiction:
                  type: string
                  maxLength: 200
                percentage:
                  type: number
                  minimum: 0
                  maximum: 100
                inclusive:
                  type: boolean
                  default: false
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              basic:
                value:
                  display_name: CA State Tax
                  jurisdiction: California
                  percentage: 7.25
                  inclusive: false
      responses:
        "201":
          description: Tax rate created successfully.
          content:
            application/json:
              schema:
                type: object
                required:
                  - tax_rate
                properties:
                  tax_rate:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/subscriptions/{id}:
    get:
      summary: Get subscription
      description: Retrieves a single subscription by ID.
      x-platform:
        kind: resource
        entity: Subscription
        affectsLedger: false
      operationId: getSubscription
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Subscription details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - subscription
                properties:
                  subscription:
                    $ref: "#/components/schemas/Subscription"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update subscription
      description: Updates subscription metadata, default payment method, or cancel-at-period-end flag.
      x-platform:
        kind: resource
        entity: Subscription
        emitsEvents:
          - subscription.updated
        affectsLedger: false
      operationId: updateSubscription
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                default_payment_method:
                  type: string
                cancel_at_period_end:
                  type: boolean
                metadata:
                  type: object
      responses:
        "200":
          description: Subscription updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - subscription
                properties:
                  subscription:
                    $ref: "#/components/schemas/Subscription"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/subscriptions/{id}/items:
    get:
      summary: List subscription items
      description: Returns paginated items for a subscription.
      x-platform:
        kind: resource
        entity: SubscriptionItem
        affectsLedger: false
      operationId: listSubscriptionItems
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated subscription items.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SubscriptionItem"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Add subscription item
      description: Adds a new price line item to an existing subscription.
      x-platform:
        kind: resource
        entity: SubscriptionItem
        emitsEvents:
          - subscription.updated
        affectsLedger: false
      operationId: addSubscriptionItem
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - price_id
              properties:
                price_id:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
                  default: 1
      responses:
        "201":
          description: Item added.
          content:
            application/json:
              schema:
                type: object
                required:
                  - subscription_item
                properties:
                  subscription_item:
                    $ref: "#/components/schemas/SubscriptionItem"
        "400":
          description: Validation error or subscription terminated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription or price not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    Subscription:
      description: A recurring billing subscription.
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        customer_id:
          type: string
        status:
          type: string
          enum:
            - incomplete
            - incomplete_expired
            - trialing
            - active
            - past_due
            - paused
            - canceled
            - unpaid
        current_period_start:
          type: string
          format: date-time
        current_period_end:
          type: string
          format: date-time
        cancel_at_period_end:
          type: boolean
        canceled_at:
          type: string
          format: date-time
          nullable: true
        ended_at:
          type: string
          format: date-time
          nullable: true
        trial_start:
          type: string
          format: date-time
          nullable: true
        trial_end:
          type: string
          format: date-time
          nullable: true
        default_payment_method:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionItem"
    SubscriptionItem:
      description: A line item on a subscription linking a price.
      type: object
      properties:
        id:
          type: string
        subscription_id:
          type: string
        price_id:
          type: string
        quantity:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...